form validation moved to utils
