+++ src/containers/Profile/ProfileForm/ProfileFormCompetencies.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { FieldArray, reduxForm, propTypes as reduxFormPropTypes } from 'redux-form';
 import { PropTypes } from 'prop-types';
 import renderCompetencies from './render/RenderCompetencies';
-import validate from './validate/validate';
+import validate from '../../../utils/validate';
 import '../ProfileStyles/Forms.css';
 
 const ProfileFormCompetencies = (props) => {
