+++ src/containers/Profile/ProfileForm/ProfileFormExamplesOfWork.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { FieldArray, reduxForm, propTypes as reduxFormPropTypes } from 'redux-form';
 import { PropTypes } from 'prop-types';
 import renderExamplesOfWork from './render/RenderExamplesOfWork';
-import validate from './validate/validate';
+import validate from '../../../utils/validate';
 
 const ProfileFormExamplesOfWork = (props) => {
   const { handleSubmit, previousPage, nextPage } = props;
