+++ src/containers/Profile/ProfileForm/ProfileFormPhotoBio.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Field, reduxForm, propTypes as reduxFormPropTypes } from 'redux-form';
 import renderField from './render/RenderField';
-import validate from './validate/validate';
+import validate from '../../../utils/validate';
 
 const ProfileFormPhotoBio = (props) => {
   const { handleSubmit, nextPage } = props;
