+++ src/containers/Profile/ProfileStyles/Forms.css
@@ -14,8 +14,12 @@
 }
 .profile-tabs {
   display: flex;
+  flex-wrap: wrap;
   justify-content: space-between;
   margin-bottom: 30px;
+}
+.profile-tabs > * {
+  flex: 1 1 160px;
 }
 .profile-tab {
   padding: 5px 15px;
