+++ src/utils/validate.js
@@ -1,11 +1,35 @@
 const validate = (values) => {
   const errors = {};
+
+  // login form
+  if (!values.email) {
+    errors.email = 'Required';
+  } else if (!/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i.test(values.email)) {
+    errors.email = 'Invalid email address';
+  }
+
   if (!values.firstName) {
-    errors.firstName = 'First name is Required';
+    errors.firstName = 'Required';
+  } else if (values.firstName.length < 2 || values.firstName.length > 20) {
+    errors.firstName = 'First name must be between 2 and 20 characters';
+  } else if (!/^[a-zA-Z ]*$/.test(values.firstName)) {
+    errors.firstName = 'First name must not contain numbers.';
   }
+
   if (!values.lastName) {
-    errors.lastName = 'Last Name is Required';
+    errors.lastName = 'Required';
+  } else if (values.lastName.length < 2 || values.lastName.length > 20) {
+    errors.lastName = 'Last name must be between 2 and 20 characters';
+  } else if (!/^[a-zA-Z ]*$/.test(values.lastName)) {
+    errors.lastName = 'Last name must not contain numbers.';
   }
+
+  if (!values.password) {
+    errors.password = 'Required';
+  } else if (values.password.length < 6) {
+    errors.password = 'Password must be longer than 6 characters.';
+  }
+  // profile form
   if (!values.bio) {
     errors.bio = 'Bio Required';
   }
